added support for .hbs
