+++ bin/preprocessor
@@ -1,0 +1,6 @@
+#!/usr/bin/env node
+
+var path = require('path');
+var fs = require('fs');
+var lib = path.join(path.dirname(fs.realpathSync(__filename)), '../lib');
+require(lib + '/main.js');
