+++ tests/file-system/app/templates/foo.handlebars
@@ -1,0 +1,1 @@
+{{outlet}}
