+++ tests/file-system/app/templates/tables/index.handlebars
@@ -1,0 +1,1 @@
+{{outlet}}
