+++ tests/file-system/app/templates/tables/other.hbs
@@ -1,0 +1,1 @@
+{{outlet}}
